New attribute * Buggies can now have individual names, no longer relying user to simply go by `buggy_id`.

+++ scripts/calcs.py
@@ -47,20 +47,20 @@ def calc_total_cost(user_choices):
         else:
             cost = "0"
         
-        total_cost += int(cost)
+        total_cost += float(cost)
 
 
     # calc tyres
     user_num_tyres = user_choices['qty_tyres']
     user_tyre_type = user_choices['tyres']
     tyre_type_cost = att_costs['tyres'][user_tyre_type]['cost']
-    total_cost += user_num_tyres * int(tyre_type_cost)
+    total_cost += user_num_tyres * float(tyre_type_cost)
 
     # calc armour, since you can only have 1 type we just add to total
     # HOWEVER, the cost changes when the num of wheels in a buggy
     # is >4. Which each wheel above 4, 10% of cost is added per wheel.
     user_armour = user_choices['armour']
-    user_num_wheels = int(user_choices['qty_wheels'])
+    user_num_wheels = float(user_choices['qty_wheels'])
 
     # if num of wheel >4, percentage addition is made but if 4, then it's simply
     # final calculation multipled by 1, so nothing changes unless it's >4. 
@@ -70,19 +70,19 @@ def calc_total_cost(user_choices):
     # adding 1 so percentage is added on top of original value, instead of simply calculating its percentage.
 
     armour_cost = att_costs['armour'][user_armour]['cost']
-    user_armour_cost = int(armour_cost) * remainder_wheels_cost
+    user_armour_cost = float(armour_cost) * remainder_wheels_cost
     total_cost += user_armour_cost
 
     # calc attack
     user_num_attacks = user_choices['qty_attacks']
     user_attack = user_choices['attack']
     cost_attack = att_costs['attack'][user_attack]['cost']
-    total_cost += user_num_attacks * int(cost_attack)
+    total_cost += user_num_attacks * float(cost_attack)
 
     # calc hasmter boosters
     user_num_hams = user_choices['hamster_booster']
     cost_of_hams = att_costs['special']['hamster_booster']['cost']
-    total_cost += user_num_hams * int(cost_of_hams)
+    total_cost += user_num_hams * float(cost_of_hams)
     
 
     ## because 'none' is not included in the 'power_type' portion of 
@@ -92,13 +92,13 @@ def calc_total_cost(user_choices):
     if user_aux_type != "none":
         user_num_aux_powers = user_choices['aux_power_units']
         cost_of_aux_type = att_costs['power_type'][user_aux_type]['cost']
-        total_cost += user_num_aux_powers * int(cost_of_aux_type)
+        total_cost += user_num_aux_powers * float(cost_of_aux_type)
 
     # calc power types
     user_power_type = user_choices['power_type']
     user_num_powers = user_choices['power_units']
     cost_of_power_type = att_costs['power_type'][user_power_type]['cost']
-    total_cost += user_num_powers * int(cost_of_power_type)
+    total_cost += user_num_powers * float(cost_of_power_type)
 
     
     return total_cost
@@ -127,13 +127,13 @@ def calc_total_mass(user_choices):
     user_num_tyres = user_choices['qty_tyres']
     user_tyre_type = user_choices['tyres']
     tyre_type_mass = att_costs['tyres'][user_tyre_type]['mass']
-    total_mass += user_num_tyres * int(tyre_type_mass)
+    total_mass += user_num_tyres * float(tyre_type_mass)
 
     # calc armour, since you can only have 1 type we just add to total
     # HOWEVER, the mass changes when the num of wheels in a buggy
     # is >4. Which each wheel above 4, 10% of mass is added per wheel.
     user_armour = user_choices['armour']
-    user_num_wheels = int(user_choices['qty_wheels'])
+    user_num_wheels = float(user_choices['qty_wheels'])
 
     # if num of wheel >4, percentage addition is made but if 4, then it's simply
     # final calculation multipled by 1, so nothing changes unless it's >4. 
@@ -143,14 +143,14 @@ def calc_total_mass(user_choices):
     # adding 1 so percentage is added on top of original value, instead of simply calculating its percentage.
 
     armour_mass = att_costs['armour'][user_armour]['mass']
-    user_armour_mass = int(armour_mass) * remainder_wheels_cost
+    user_armour_mass = float(armour_mass) * remainder_wheels_cost
     total_mass += user_armour_mass
 
     # calc attack
     user_num_attacks = user_choices['qty_attacks']
     user_attack = user_choices['attack']
     mass_attack = att_costs['attack'][user_attack]['mass']
-    total_mass += user_num_attacks * int(mass_attack)
+    total_mass += user_num_attacks * float(mass_attack)
 
     ## because 'none' is not included in the 'power_type' portion of 
     ## of the JSON, but still an option for the aux power type, 
@@ -160,13 +160,13 @@ def calc_total_mass(user_choices):
     if user_aux_type != "none":
         user_num_aux_powers = user_choices['aux_power_units']
         mass_of_aux_type = att_costs['power_type'][user_aux_type]['mass']
-        total_mass += user_num_aux_powers * int(mass_of_aux_type)
+        total_mass += user_num_aux_powers * float(mass_of_aux_type)
 
     # calc power types
     user_power_type = user_choices['power_type']
     user_num_powers = user_choices['power_units']
     mass_of_power_type = att_costs['power_type'][user_power_type]['mass']
-    total_mass += user_num_powers * int(mass_of_power_type)
+    total_mass += user_num_powers * float(mass_of_power_type)
 
 
     return total_mass
